--- config-local-kaldi.yml
@@ -1,46 +1,0 @@
-# Important for understanding DANE configs: https://github.com/CLARIAH/DANE/blob/main/DANE/config.py
-
-# Important note:
-# FIRST the home dir config is applied (~/.DANE/config.yml),
-# THEN the base_config.yml will overwrite anything,
-# THEN the local config.yml
-
-#Note: For local testing, copy this file to config.yml (in main dir of this repo)
-RABBITMQ:
-    HOST: 'dane-rabbitmq-api.default.svc.cluster.local'
-    PORT: 5672
-    EXCHANGE: 'DANE-exchange'
-    RESPONSE_QUEUE: 'DANE-response-queue'
-    USER: 'guest' # change this for production mode
-    PASSWORD: 'guest' # change this for production mode
-ELASTICSEARCH:
-    HOST: ['elasticsearch']
-    PORT: 9200
-    USER: '' # change this for production mode
-    PASSWORD: '' # change this for production mode
-    SCHEME: 'http'
-    INDEX: 'dane-index-k8s-asr'
-LOCAL_KALDI:  # TODO use this in worker.py
-    ASR_PACKAGE_NAME: 'asr-features.tar.gz'
-    ASR_WORD_JSON_FILE: 'words.json'
-    # (as module) kaldi NL base dir and the decoder script therein
-    KALDI_NL_DIR: '/opt/Kaldi_NL'
-    KALDI_NL_DECODER: 'decode_OH.sh'  # 'decode.sh'
-    KALDI_NL_MODEL_DIR: '/models'  # language models will be downloaded here (should be /mnt/dane-fs/models when kaldi_nl supports this)
-    KALDI_NL_MODEL_FETCHER: 'entrypoint.sh'
-FILE_SYSTEM:
-    BASE_MOUNT: 'mount' #'mount' when running locally
-    INPUT_DIR: 'input-files'
-    OUTPUT_DIR: 'output-files/asr-output'
-PATHS:
-    TEMP_FOLDER: "mount/input-files"
-    OUT_FOLDER: "mount/output-files"
-INPUT:
-    DELETE_ON_COMPLETION: True # this param is optional, will be defaulted to False if absent
-OUTPUT:
-    DELETE_ON_COMPLETION: True # this param is optional, will be defaulted to False if absent
-    TRANSFER_ON_COMPLETION: True
-    S3_ENDPOINT_URL: "http://localhost:9000"
-    S3_BUCKET: "test-bucket"  # bucket reserved for 1 type of output
-    S3_FOLDER_IN_BUCKET: "assets"  # folder within the bucket
-DANE_DEPENDENCIES: ['DOWNLOAD'] # this param is optional, will be defaulted to [] if absent
